+++ src/CAGMakers.js
@@ -2,7 +2,6 @@ const CAG = require('./CAG')
 const Side = require('./math/Side')
 const Vector2D = require('./math/Vector2')
 const Vertex = require('./math/Vertex2')
-const {areaEPS} = require('./constants')
 
 /** Reconstruct a CAG from an object with identical property names.
  * @param {Object} obj - anonymous object, typically from JSON
@@ -12,54 +11,9 @@ const fromObject = function (obj) {
   let sides = obj.sides.map(function (s) {
     return Side.fromObject(s)
   })
-  let cag = fromSides(sides)
+  let cag = CAG.fromSides(sides)
   cag.isCanonicalized = obj.isCanonicalized
   return cag
-}
-
-/** Construct a CAG from a list of `Side` instances.
- * @param {Side[]} sides - list of sides
- * @returns {CAG} new CAG object
- */
-const fromSides = function (sides) {
-  let cag = new CAG()
-  cag.sides = sides
-  return cag
-}
-
-/** Construct a CAG from a list of points (a polygon).
- * The rotation direction of the points is not relevant.
- * The points can define a convex or a concave polygon.
- * The polygon must not self intersect.
- * @param {points[]} points - list of points in 2D space
- * @returns {CAG} new CAG object
- */
-const fromPoints = function (points) {
-  let numpoints = points.length
-  if (numpoints < 3) throw new Error('CAG shape needs at least 3 points')
-  let sides = []
-  let prevpoint = new Vector2D(points[numpoints - 1])
-  let prevvertex = new Vertex(prevpoint)
-  points.map(function (p) {
-    let point = new Vector2D(p)
-    let vertex = new Vertex(point)
-    let side = new Side(prevvertex, vertex)
-    sides.push(side)
-    prevvertex = vertex
-  })
-  let result = fromSides(sides)
-  if (result.isSelfIntersecting()) {
-    throw new Error('Polygon is self intersecting!')
-  }
-  let area = result.area()
-  if (Math.abs(area) < areaEPS) {
-    throw new Error('Degenerate polygon!')
-  }
-  if (area < 0) {
-    result = result.flipped()
-  }
-  result = result.canonicalized()
-  return result
 }
 
 /** Construct a CAG from a list of points (a polygon).
@@ -81,7 +35,7 @@ const fromPointsNoCheck = function (points) {
     sides.push(side)
     prevvertex = vertex
   })
-  return fromSides(sides)
+  return CAG.fromSides(sides)
 }
 
 // Converts a CSG to a  The CSG must consist of polygons with only z coordinates +1 and -1
@@ -93,13 +47,11 @@ const fromFakeCSG = function (csg) {
         .filter(function (s) {
           return s !== null
         })
-  return fromSides(sides)
+  return CAG.fromSides(sides)
 }
 
 module.exports = {
   fromObject,
-  fromSides,
-  fromPoints,
   fromPointsNoCheck,
   fromFakeCSG
 }
